backend/requirements.txt
facebook-sdk
flask
flask-cache